backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# CORS support
fastapi-cors==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
requests==2.31.0
httpx==0.25.2

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.15.2

# Quantum computing libraries
qiskit==0.45.1
qiskit-terra==0.45.1
pennylane==0.34.0
pennylane-lightning==0.36.0

# Geocoding
geopy==2.3.0
geographiclib==2.0

# Visualization (for quantum circuits)
plotly==5.18.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1 